locking of nodes

+++ smart-contracts/wasm-chain-integration/src/v1/types.rs
@@ -495,6 +495,7 @@ pub struct InstanceState<'a, BackingStore> {
     iterators:          Vec<Option<trie::Iterator>>,
     /// Opaque pointer to the state of the instance in consensus.
     state_trie:         trie::StateTrie<'a>,
+    iterator_roots:     std::collections::BTreeMap<usize, Vec<u8>>,
 }
 
 /// first bit is ignored, the next 31 indicate a generation,
@@ -614,10 +615,15 @@ impl<'a, BackingStore: trie::FlatLoadable> InstanceState<'a, BackingStore> {
             state_trie: state.state.lock().unwrap(),
             iterators: Vec::new(),
             entry_mapping: Vec::new(),
+            iterator_roots: std::collections::BTreeMap::new(),
         }
     }
 
     pub fn lookup_entry(&mut self, key: &[u8]) -> InstanceStateEntryOption {
+        // check that the node is not part of an iterator
+        if is_key_locked(key, &self.iterator_roots) {
+            return InstanceStateEntryOption::new(None);
+        }
         if let Some(id) = self.state_trie.get_entry(&mut self.backing_store, key) {
             let idx = self.entry_mapping.len();
             self.entry_mapping.push(Some(EntryWithKey {
@@ -648,6 +654,10 @@ impl<'a, BackingStore: trie::FlatLoadable> InstanceState<'a, BackingStore> {
         ensure!(gen == self.current_generation, "Incorrect entry id generation.");
         let entry = if let Some(entry) = self.entry_mapping.get_mut(idx) {
             if let Some(entry) = std::mem::take(entry) {
+                // check that the node is not part of an iterator
+                if is_key_locked(&entry.key, &self.iterator_roots) {
+                    return Ok(0);
+                }
                 entry
             } else {
                 return Ok(0);
@@ -663,6 +673,10 @@ impl<'a, BackingStore: trie::FlatLoadable> InstanceState<'a, BackingStore> {
     }
 
     pub fn delete_prefix(&mut self, key: &[u8]) -> StateResult<u32> {
+        if is_key_locked(key, &self.iterator_roots) {
+            return Ok(0);
+        }
+
         if self.state_trie.delete_prefix(&mut self.backing_store, key).is_some() {
             Ok(1)
         } else {
@@ -674,6 +688,7 @@ impl<'a, BackingStore: trie::FlatLoadable> InstanceState<'a, BackingStore> {
         if let Some(iter) = self.state_trie.iter(&mut self.backing_store, prefix) {
             let iter_id = self.iterators.len();
             self.iterators.push(Some(iter));
+            self.iterator_roots.insert(iter_id, prefix.to_vec());
             InstanceStateIteratorOption::new(Some((self.current_generation, iter_id)))
         } else {
             InstanceStateIteratorOption::new(None)
@@ -707,9 +722,9 @@ impl<'a, BackingStore: trie::FlatLoadable> InstanceState<'a, BackingStore> {
         ensure!(gen == self.current_generation, "Incorrect iterator generation.");
         match self.iterators.get_mut(idx) {
             Some(iter) => match iter {
-                Some(existing_iter) => {
+                Some(_) => {
                     // Unlock the nodes associated with this iterator.
-                    self.state_trie.iter_delete(&mut self.backing_store, existing_iter);
+                    self.iterator_roots.remove(&idx);
                     // Finally we remove the iterator by setting it to `None`.
                     *iter = None;
                     Ok(1)
@@ -870,4 +885,25 @@ impl<'a, BackingStore: trie::FlatLoadable> InstanceState<'a, BackingStore> {
             bail!("Invalid entry.");
         }
     }
+}
+
+fn is_key_locked(key: &[u8], iterator_roots: &std::collections::BTreeMap<usize, Vec<u8>>) -> bool {
+    for iter_root in iterator_roots.values() {
+        if key.starts_with(iter_root) {
+            return true;
+        }
+    }
+    false
+}
+
+#[test]
+fn test_is_key_locked() {
+    let mut iterator_roots = std::collections::BTreeMap::new();
+    iterator_roots.insert(0, vec![0, 1]);
+    iterator_roots.insert(1, vec![0, 0, 1]);
+    assert!(!is_key_locked(&[0], &iterator_roots), "Subtree should not be locked");
+    assert!(!is_key_locked(&[0, 0], &iterator_roots), "Subtree should not be locked");
+    assert!(!is_key_locked(&[0, 0, 0], &iterator_roots), "Subtree should not be locked");
+    assert!(is_key_locked(&[0, 1], &iterator_roots), "Subtree should be locked");
+    assert!(is_key_locked(&[0, 1, 0], &iterator_roots), "Subtree should be locked");
 }
